backend/app/requirements.txt
fastapi
uvicorn
psycopg2-binary
redis
prometheus_client
